refactor: decoupling code a bit
