+++ apps/namadillo/src/App/Common/FixedWarningBanner.tsx
@@ -1,46 +1,13 @@
-import { indexerHeartbeatAtom } from "atoms/settings";
-import { useNamadaKeychain } from "hooks/useNamadaKeychain";
-import { useAtomValue } from "jotai";
-import { ReactNode, useEffect, useState } from "react";
+import { ReactNode } from "react";
 import { IoWarning } from "react-icons/io5";
-import {
-  checkIndexerCompatibilityErrors,
-  checkKeychainCompatibilityError,
-} from "utils/compatibility";
 
-export const FixedWarningBanner = (): JSX.Element => {
-  const [errorMessage, setErrorMessage] = useState<ReactNode | undefined>();
-  const indexerHealth = useAtomValue(indexerHeartbeatAtom);
-  const keychain = useNamadaKeychain();
+type FixedWarningBannerProps = {
+  errorMessage: ReactNode;
+};
 
-  const verifyKeychainVersion = async (): Promise<void> => {
-    const namadaKeychain = await keychain.namadaKeychain.get();
-    if (namadaKeychain) {
-      const version = namadaKeychain.version();
-      const versionErrorMessage = checkKeychainCompatibilityError(version);
-      if (versionErrorMessage) {
-        setErrorMessage(versionErrorMessage);
-      }
-    }
-  };
-
-  const verifyIndexerVersion = async (): Promise<void> => {
-    const versionErrorMessage = checkIndexerCompatibilityErrors(
-      indexerHealth.data?.version || ""
-    );
-    if (versionErrorMessage) {
-      setErrorMessage(versionErrorMessage);
-    }
-  };
-
-  useEffect(() => {
-    verifyKeychainVersion();
-  }, [keychain]);
-
-  useEffect(() => {
-    indexerHealth.isSuccess && verifyIndexerVersion();
-  }, [indexerHealth]);
-
+export const FixedWarningBanner = ({
+  errorMessage,
+}: FixedWarningBannerProps): JSX.Element => {
   if (!errorMessage) return <></>;
 
   return (
