+++ apps/namadillo/src/App/Layout/AppLayout.tsx
@@ -1,4 +1,5 @@
 import { FixedWarningBanner } from "App/Common/FixedWarningBanner";
+import { useCompatibilityErrors } from "hooks/useCompatibilityErrors";
 import { ReactNode, useState } from "react";
 import { twMerge } from "tailwind-merge";
 import { AppHeader } from "./AppHeader";
@@ -11,6 +12,7 @@ export const AppLayout = ({
   children: ReactNode;
 }): JSX.Element => {
   const [displayNavigation, setDisplayNavigation] = useState(false);
+  const compatibilityErrors = useCompatibilityErrors();
 
   return (
     <div className="custom-container pb-2">
@@ -43,7 +45,7 @@ export const AppLayout = ({
         </aside>
         <main className="min-h-full">{children}</main>
       </div>
-      <FixedWarningBanner />
+      <FixedWarningBanner errorMessage={compatibilityErrors} />
     </div>
   );
 };
